# Fix: no-trailing-spaces now handles skipBlankLines (fixes #2575)