--- a/tests/lib/rules/no-trailing-spaces.js
+++ b/tests/lib/rules/no-trailing-spaces.js
@@ -1,6 +1,7 @@
 /**
  * @fileoverview Disallow trailing spaces at the end of lines.
  * @author Nodeca Team <https://github.com/nodeca>
+ * @copyright 2015 Patrick McElhaney
  */
 "use strict";
 
@@ -144,6 +145,18 @@ eslintTester.addRuleTest("lib/rules/no-trailing-spaces", {
                 type: "Program"
             }],
             options: [{}]
+        },
+        {
+            code: "var a = 'bar';  \n \n\t",
+            errors: [{
+                message: "Trailing spaces not allowed.",
+                type: "Program",
+                line: 1,
+                column: 16 // there are invalid spaces in columns 15 and 16
+            }],
+            options: [{
+                skipBlankLines: true
+            }]
         }
     ]