This does potentially two regexp tests per line. With this change, the performance does not seem to be impacted in a statistically significant way.

With: 
```
node Makefile.js perf
CPU Speed is 2200 with multiplier 7500000
Performance Run #1:  2177.437605ms
Performance Run #2:  2095.757845ms
Performance Run #3:  2085.389922ms
Performance Run #4:  2096.332165ms
Performance Run #5:  2082.197974ms
Performance budget ok:  2095.757845ms (limit: 3409.090909090909ms)
```
Without:
```
node Makefile.js perf
CPU Speed is 2200 with multiplier 7500000
Performance Run #1:  2085.415068ms
Performance Run #2:  2181.624126ms
Performance Run #3:  2082.826633ms
Performance Run #4:  2136.809016ms
Performance Run #5:  2126.298463ms
Performance budget ok:  2126.298463ms (limit: 3409.090909090909ms)
```