@gcochard 2015-06-06
Comments addressed. Good to go.